feat: moved technology-card to shared components, added new fields only visible to admin
