--- frontend/src/app/modules/home/components/technology-card/technology-card.component.html
@@ -1,69 +1,0 @@
-<div
-  class="w-full block p-4 mb-4 border border-gray-900 rounded-lg shadow bg-gray-700 border-gray-700"
->
-  <h5
-    class="mb-2 text-2xl font-bold tracking-tight text-gray-900 text-white"
-  >
-    {{ technology.name }}
-  </h5>
-  <p class="font-normal text-white text-white">
-    {{ technology.description }}
-  </p>
-  <div class="mt-3" id="accordion-open" data-accordion="open">
-    <h2 id="accordion-open-heading-1">
-      <button
-        type="button"
-        class="flex items-center justify-between w-full p-5 font-medium text-white bg-gray-800 border border-gray-900 rounded-t-xl focus:ring-4 focus:ring-gray-700  hover:bg-gray-800 gap-3"
-        [class.bg-gray-800] ="isRingDescriptionOpen"
-        data-accordion-target="#accordion-open-body-1"
-        aria-expanded="true"
-        aria-controls="accordion-open-body-1"
-        (click)="toggleRingDescrition()"
-      >
-        <span class="flex items-center"
-          ><svg
-            class="w-5 h-5 me-2 shrink-0"
-            fill="currentColor"
-            viewBox="0 0 20 20"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              fill-rule="evenodd"
-              d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-8-3a1 1 0 00-.867.5 1 1 0 11-1.731-1A3 3 0 0113 8a3.001 3.001 0 01-2 2.83V11a1 1 0 11-2 0v-1a1 1 0 011-1 1 1 0 100-2zm0 8a1 1 0 100-2 1 1 0 000 2z"
-              clip-rule="evenodd"
-            ></path>
-          </svg>
-          Why on this ring?</span
-        >
-        <svg
-          data-accordion-icon
-          class="w-3 h-3 shrink-0"
-          [class.rotate-180]="isRingDescriptionOpen"
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 10 6"
-        >
-          <path
-            stroke="currentColor"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="2"
-            d="M9 5 5 1 1 5"
-          />
-        </svg>
-      </button>
-    </h2>
-    <div *ngIf="isRingDescriptionOpen"
-      id="accordion-open-body-1"
-      aria-labelledby="accordion-open-heading-1"
-    >
-      <div
-        class="p-5 bg-gray-800 border border-gray-700"
-      >
-        <p class="mb-2 text-white">
-            {{technology.descriptionCategorization}}
-        </p>
-      </div>
-    </div>
-  </div>
-</div>
