+++ frontend/src/app/shared/components/technology-card/technology-card.component.html
@@ -1,0 +1,66 @@
+<div class="w-full block p-4 mb-4 border border-gray-900 rounded-lg shadow bg-gray-700 border-gray-700">
+  <h5 class="mb-2 text-2xl font-bold tracking-tight text-gray-900 text-white">
+    {{ technology.name }}
+  </h5>
+  <p class="font-normal text-white text-white">
+    {{ technology.description }}
+  </p>
+  <div class="mt-3">
+    <div *ngIf="technology.descriptionCategorization;else emptyDescriptionCategorization" id="accordion-open"
+      data-accordion="open">
+      <h2 id="accordion-open-heading-1">
+        <button type="button"
+          class="flex items-center justify-between w-full p-5 font-medium text-white bg-gray-800 border border-gray-900 rounded-t-xl focus:ring-4 focus:ring-gray-700  hover:bg-gray-800 gap-3"
+          [class.bg-gray-800]="isRingDescriptionOpen" data-accordion-target="#accordion-open-body-1"
+          aria-expanded="true" aria-controls="accordion-open-body-1" (click)="toggleRingDescrition()">
+          <span class="flex items-center"><svg class="w-5 h-5 me-2 shrink-0" fill="currentColor" viewBox="0 0 20 20"
+              xmlns="http://www.w3.org/2000/svg">
+              <path fill-rule="evenodd"
+                d="M18 10a8 8 0 11-16 0 8 8 0 0116 0zm-8-3a1 1 0 00-.867.5 1 1 0 11-1.731-1A3 3 0 0113 8a3.001 3.001 0 01-2 2.83V11a1 1 0 11-2 0v-1a1 1 0 011-1 1 1 0 100-2zm0 8a1 1 0 100-2 1 1 0 000 2z"
+                clip-rule="evenodd"></path>
+            </svg>
+            Why on this ring?</span>
+          <svg data-accordion-icon class="w-3 h-3 shrink-0" [class.rotate-180]="isRingDescriptionOpen"
+            xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 10 6">
+            <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+              d="M9 5 5 1 1 5" />
+          </svg>
+        </button>
+      </h2>
+      <div *ngIf="isRingDescriptionOpen" id="accordion-open-body-1" aria-labelledby="accordion-open-heading-1">
+        <div class="p-5 bg-gray-800 border border-gray-700">
+          <p class="mb-2 text-white">
+            {{technology.descriptionCategorization}}
+          </p>
+        </div>
+      </div>
+    </div>
+    <ng-template #emptyDescriptionCategorization>
+      <app-alert alertType="info" message="A categorization description is missing." />
+    </ng-template>
+  </div>
+  <div *ngIf="technology._id" class="my-4">
+    <div>
+      <div
+        class="bg-green-700 bg-opacity-25 text-white border-2 border-green-800 text-xs font-medium me-2 px-2.5 py-2.5 rounded mb-3">
+        {{technology.creator?.userName}} on {{technology.createdAt}}
+      </div>
+      <div *ngIf="!technology.published; else publicationInfos"
+        class="bg-yellow-700 bg-opacity-25 text-white border-2 border-yellow-800 text-xs font-medium me-2 px-2.5 py-2.5 rounded mb-3">
+        Not Published yet.
+      </div>
+      <ng-template #publicationInfos>
+        <div
+          class="bg-blue-700 bg-opacity-25 text-white border-2 border-blue-800 text-xs font-medium me-2 px-2.5 py-2.5 rounded ">
+          {{technology.publisher?.userName}} on {{technology.publishedAt}}
+        </div>
+      </ng-template>
+    </div>
+    <div>
+      <a *ngIf="!technology.published" [routerLink]="['/admin/technology', technology._id, 'publish']"
+        class="text-gray-900 dark:text-white hover:underline">
+        Publish
+      </a>
+    </div>
+  </div>
+</div>
