+++ frontend/src/app/shared/components/technology-card/technology-card.component.spec.ts
